// src/types/resources/Slot.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TInstant } from '../simpleTypes/Instant';

export type TSlot = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  serviceCategory?: TCodeableConcept[];
  serviceType?: TCodeableConcept[];
  specialty?: TCodeableConcept[];
  appointmentType?: TCodeableConcept;
  schedule: TReference;
  status: string;
  start: TInstant;
  end: TInstant;
  overbooked?: boolean;
  comment?: string;
};